Added logging statements backend
Scan start, restart time and sleep being logged to the console for the user during runtime.

Removed print statements from the dbFetch.fetch("") function.

Added a timeNow in File.py to log to the database fileScannedAt field  for multiple entries. Cleanup needs to happen on the database end to reduce multiple occurrences now.