stable beta!
comment documentation added to directoryScraper.py

if clause added to avoid notifying the user of the initial login scan which occurs every launch.

notifications added to tell the user when a scan is finished

sec variable added to track refresh interval in the future for the auto mode - currently commented out.